# OrderManagement/src/main/java/lab25/ordermanagement/articles/dto/ArticleDTO.java
package lab25.ordermanagement.articles.dto;

import lab25.ordermanagement.articles.enums.ArticleType;

public record ArticleDTO(
        String id,
        ArticleType type,
        String name,
        String articleNumber,
        String description
) {
}