# OrderManagementFrontend/openapi.yaml
openapi: 3.1.0
info:
  title: Order Management API
  version: "1.0"
servers:
- url: http://localhost:8092/ordermanagement/api
  description: Generated server url
paths:
  /articles/{articleId}:
    get:
      tags:
      - article-controller
      operationId: getArticleById
      parameters:
      - name: articleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDTO"
    post:
      tags:
      - article-controller
      operationId: updateArticle
      parameters:
      - name: articleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdateDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDTO"
  /articles:
    get:
      tags:
      - article-controller
      operationId: listArticles
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageArticleDTO"
components:
  schemas:
    ArticleUpdateDTO:
      type: object
      properties:
        type:
          type: string
          enum:
          - PHYSICAL_PRODUCT
          - DIGITAL_PRODUCT
        name:
          type: string
        articleNumber:
          type: string
        description:
          type: string
    ArticleDTO:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum:
          - PHYSICAL_PRODUCT
          - DIGITAL_PRODUCT
        name:
          type: string
        articleNumber:
          type: string
        description:
          type: string
    PageArticleDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ArticleDTO"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean