# OrderManagementFrontend/src/app/generated/ordermanagement/model/articleDTO.ts
/**
 * Order Management API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ArticleDTO { 
    id?: string;
    type?: ArticleDTO.TypeEnum;
    name?: string;
    articleNumber?: string;
    description?: string;
}
export namespace ArticleDTO {
    export const TypeEnum = {
        PhysicalProduct: 'PHYSICAL_PRODUCT',
        DigitalProduct: 'DIGITAL_PRODUCT'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}