# OrderManagementFrontend/src/app/generated/ordermanagement/model/articleUpdateDTO.ts
/**
 * Order Management API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ArticleUpdateDTO { 
    type?: ArticleUpdateDTO.TypeEnum;
    name?: string;
    articleNumber?: string;
    description?: string;
}
export namespace ArticleUpdateDTO {
    export const TypeEnum = {
        PhysicalProduct: 'PHYSICAL_PRODUCT',
        DigitalProduct: 'DIGITAL_PRODUCT'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}